+++ CHANGELOG.md
@@ -2,6 +2,8 @@
 This project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
 ## Unreleased
+### Added
+- Boolean options may now be rewritten into strings using `rewrite`.
 
 ## 0.6.4 (2022-09-27)
 ### Added